finished with visualize_pointclouds
